add hardhat setCommitteeTask
