+++ .env.example
@@ -2,7 +2,3 @@ MNEMONIC="test test test test test test test test test test test junk"
 INFURA_PROJECT_ID=""
 ETHERSCAN_API_KEY=""
 POLYGONSCAN_API_KEY=""
-
-# for setup-committee.js
-DAC_URL="http://cdk-validium-data-availability:8444"
-DAC_MEMBER_ADDRESS=""
